Python: Add support for use-system-flag directive
We attempt to install dependencies from pip to a custom location using
the --target flag.  This flag is broken in Debian/Ubuntu due to a
distro-specific --user install default in versions prior to 18.1.1
(including the Stretch version).

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=830892

This behavior can be switched off by specifying the --system flag in the
install command. This patch adds support for a use-system-flag directive
that allows doing so.

Bug: T227919
Change-Id: Ifacc3d96be3a89135187e46450d524a7d083f885